+++ app/src/test/java/com/plusorminustwo/postmark/ui/thread/PinnedThreadTest.kt
@@ -116,6 +116,8 @@ class PinnedThreadTest {
 
         override suspend fun insert(thread: ThreadEntity) {}
         override suspend fun insertAll(threads: List<ThreadEntity>) {}
+        override suspend fun insertIgnore(thread: ThreadEntity) {}
+        override suspend fun insertAllIgnore(threads: List<ThreadEntity>) {}
         override suspend fun update(thread: ThreadEntity) {}
         override suspend fun delete(thread: ThreadEntity) {}
         override suspend fun updateBackupPolicy(threadId: Long, policy: BackupPolicy) {}